selected car info

+++ lib/models/account.dart
@@ -9,7 +9,7 @@ class Account {
   final String password;
   final String email;
   final String name;
-  late int power;
+  late double power;
   late double todayMiningPer = 0.0;
   late double todayMiningXPer = 0.0;
